requirements.txt
-f /usr/share/pip-wheels
Django==1.11.3
django-colorfield==0.1.14
django-crispy-forms==1.6.0
django-froala-editor==2.6.4
django-markdown==0.8.4
django-markdown-deux==1.0.5
django-tinymce==2.6.0
Markdown==2.6.8
markdown2==2.3.4
olefile==0.44
Pillow==4.2.1
pytz==2017.2
whitenoise==3.3.0
